requirements.txt
absl-py==0.11.0
asgiref==3.4.1
asn1crypto==0.24.0
astor==0.8.1
bokeh==2.3.3
click==8.0.3
cloudpickle==2.0.0
contextvars==2.4
cryptography==2.1.4
cycler==0.11.0
dask==2021.3.0
distributed==2021.3.0
Django==3.2.12
fsspec==2021.11.1
gast==0.2.2
geographiclib==1.52
geopy==2.2.0
google-pasta==0.2.0
grpcio==1.34.0
h5py==2.10.0
hdf5storage==0.1.18
HeapDict==1.0.1
idna==2.6
image==1.5.33
immutables==0.16
importlib-metadata==3.3.0
Jinja2==3.0.3
joblib==1.1.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
locket==0.2.1
Markdown==3.3.3
MarkupSafe==2.0.1
matplotlib==3.3.4
msgpack==1.0.3
numpy==1.19.5
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.5
partd==1.2.0
Pillow==8.4.0
protobuf==3.14.0
psutil==5.8.0
pycrypto==2.6.1
pygobject==3.26.1
pyparsing==3.0.6
PyQt5==5.9.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.8.2
pytz==2021.3
pyxdg==0.25
PyYAML==6.0
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
SecretStorage==2.3.1
sip==4.19.8
six==1.11.0
sklearn==0.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tblib==1.7.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.5
termcolor==1.1.0
threadpoolctl==3.0.0
toolz==0.11.2
tornado==6.1
typing-extensions==4.1.1
Werkzeug==1.0.1
wrapt==1.12.1
zict==2.0.0
zipp==3.4.0
